+++ internal/search/backend/indexers_test.go
@@ -178,11 +178,6 @@ func TestFindEndpoint(t *testing.T) {
 		hostname:  "foo.internal",
 		endpoints: []string{"foo.internal:80", "bar.internal:80"},
 		want:      "foo.internal:80",
-	}, {
-		name:      "exact without hostname",
-		hostname:  "foo",
-		endpoints: []string{"indexed-search:6070"},
-		want:      "indexed-search:6070",
 	}}
 
 	for _, tc := range cases {