+++ CHANGELOG.md
@@ -20,7 +20,7 @@ All notable changes to Sourcegraph are documented in this file.
 - Added a button "Reindex now" to the index status page. Admins can now force an immediate reindex of a repository. [#45533](https://github.com/sourcegraph/sourcegraph/pull/45533)
 - Added an option "Unlock user" to the actions dropdown on the Site Admin Users page. Admins can unlock user accounts that wer locked after too many sign-in attempts. [#45650](https://github.com/sourcegraph/sourcegraph/pull/45650)
 - Templates for certain emails sent by Sourcegraph are now configurable via `email.templates` in site configuration. [#45671](https://github.com/sourcegraph/sourcegraph/pull/45671)
-- Endpoints for `Gitserver`, `Indexed-search`, `Searcher`, and `Symbols` are generated automatically when the correct environment variables (`GITSERVER_REPLICA_COUNT`, `INDEXED_SEARCH_REPLICA_COUNT`, `SEARCHER_REPLICA_COUNT`, `SYMBOLS_REPLICA_COUNT`) are provided in Docker-Compose deployments
+- Endpoints for `Gitserver`, `Indexed-search`, `Searcher`, and `Symbols` are generated automatically when the correct environment variables (`GITSERVER_REPLICA_COUNT`, `INDEXED_SEARCH_REPLICA_COUNT`, `SEARCHER_REPLICA_COUNT`, `SYMBOLS_REPLICA_COUNT`) are provided in Docker-Compose and Kubernetes deployments. [#45862](https://github.com/sourcegraph/sourcegraph/pull/45862)
 
 ### Changed